Reinstate pre-1979 data to output
When filtering on incomplete sampling, plot 24 did not exist before 1979 and is listed in the trapping table as Sampled = 0. Because RodentAbundance filtered out all periods with a 0, we lost all data before 1979. I put in a slightly awkward hack. This allows the pre-1979 data through, but we aren't warning anyone that we're a plot light in whatever category plot 24 goes into.