Allow user to specify data location, including repo
Since not everyone is going to want to store their data in the ~/ path, I altered the function to allow the user to tell the function where to look for the data. I also added a way for the user to specify that they would like the current master repo data.