restricting the R code in vignettes
now including an environmental variable portalr_vignette_build (logical TRUE or FALSE) that defines if the R code chunks are run or not in the vignettes. the default set up by the project-level .Renviron file is TRUE, but the .Renviron file is not included in the R build (i.e. it's listed in the .Rbuildignore) so the vignettes should be build without the code chunks running during the Rbuild on CRAN (which only tests to see if the vignettes can be built, but does not save the output)